chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-css-sheet-in-shadow-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE ERROR: line 7: Uncaught TypeError: Cannot read property 'querySelector' of undefined
This is a testharness.js-based test.
FAIL import-css-sheet-in-shadow Uncaught TypeError: Cannot read property 'querySelector' of undefined
Harness: the test ran to completion.
